requirements.txt
torch==2.2.0
torchvision==0.17.0
editdistance==0.8.1
torchaudio==2.2.0
pyctcdecode==0.5.0
soundfile
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
wget
kenlm
gdown
shutil
OmegaConf
json

black
isort
pre-commit
